Create README.md
I'm 30, from Minnesota, USA. I don't have any background in computers or programing but am in the process of learning. I'm a perfectionist when it comes to my work/job, almost a curse and a blessing. It's helped me get to where I am now and hopefully can help me reach my goal. Happy to become apart of the coding world!